# infra/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "GODS Platform Core",
  "description": "Sovereign AI governance backend for the G.O.D.S ecosystem.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/auth/register": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Register",
    "operationId": "register_auth_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegisterReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_auth_login_post",
    "requestBody": {
     "content": {
      "application/x-www-form-urlencoded": {
       "schema": {
        "$ref": "#/components/schemas/Body_login_auth_login_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/auth/me": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Me",
    "operationId": "me_auth_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
...
  "/decisions": {
   "get": {
    "tags": [
     "UDOC decisions"
    ],
    "summary": "List Decisions",
    "operationId": "list_decisions_decisions_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "UDOC decisions"
    ],
    "summary": "Decide",
    "operationId": "decide_decisions_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DecisionReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
...
  "/oversight/cases": {
   "get": {
    "tags": [
     "UDOC oversight"
    ],
    "summary": "List Cases",
    "operationId": "list_cases_oversight_cases_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "UDOC oversight"
    ],
    "summary": "Open Case",
    "operationId": "open_case_oversight_cases_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CaseReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/oversight/cases/{case_ref}/resolve": {
   "post": {
    "tags": [
     "UDOC oversight"
    ],
    "summary": "Resolve",
    "operationId": "resolve_oversight_cases__case_ref__resolve_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "case_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Ref"
      }
     },
     {
      "name": "resolution",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Resolution"
      }
     },
     {
      "name": "override",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Override"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/analytics/{division}/kpis": {
   "get": {
    "tags": [
     "UDOC analytics"
    ],
    "summary": "Kpis",
    "operationId": "kpis_analytics__division__kpis_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "division",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Division"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/analytics/{division}/totals": {
   "get": {
    "tags": [
     "UDOC analytics"
    ],
    "summary": "Totals",
    "operationId": "totals_analytics__division__totals_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "division",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Division"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/analytics/{division}/timeseries": {
   "get": {
    "tags": [
     "UDOC analytics"
    ],
    "summary": "Timeseries",
    "operationId": "timeseries_analytics__division__timeseries_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "division",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Division"
      }
     },
     {
      "name": "metric",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Metric"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/analytics/{division}/records": {
   "get": {
    "tags": [
     "UDOC analytics"
    ],
    "summary": "Records",
    "operationId": "records_analytics__division__records_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "division",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Division"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_login_auth_login_post": {
    "properties": {
     "grant_type": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^password$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grant Type"
     },
     "username": {
      "type": "string",
      "title": "Username"
     },
     "password": {
      "type": "string",
      "format": "password",
      "title": "Password"
     },
     "scope": {
      "type": "string",
      "title": "Scope",
      "default": ""
     },
     "client_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Id"
     },
     "client_secret": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "format": "password",
      "title": "Client Secret"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "Body_login_auth_login_post"
   },
   "CaseReq": {
    "properties": {
     "model_id": {
      "type": "string",
      "title": "Model Id"
     },
     "reason": {
      "type": "string",
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "model_id",
     "reason"
    ],
    "title": "CaseReq"
   },
   "DecisionReq": {
    "properties": {
     "model_id": {
      "type": "string",
      "title": "Model Id"
     },
     "risk_tier": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Risk Tier"
     },
     "raw_confidence": {
      "type": "number",
      "title": "Raw Confidence",
      "default": 0.9
     },
     "compliance": {
      "type": "number",
      "title": "Compliance",
      "default": 1.0
     },
     "priv_favorable": {
      "type": "integer",
      "title": "Priv Favorable",
      "default": 480
     },
     "priv_total": {
      "type": "integer",
      "title": "Priv Total",
      "default": 1000
     },
     "unpriv_favorable": {
      "type": "integer",
      "title": "Unpriv Favorable",
      "default": 470
     },
     "unpriv_total": {
      "type": "integer",
      "title": "Unpriv Total",
      "default": 1000
     },
     "ecs": {
      "type": "number",
      "title": "Ecs",
      "default": 0.75
     },
     "bgp": {
      "type": "number",
      "title": "Bgp",
      "default": 1.0
     },
     "traceroute": {
      "type": "number",
      "title": "Traceroute",
      "default": 1.0
     },
     "dnssec": {
      "type": "number",
      "title": "Dnssec",
      "default": 1.0
     },
     "storage": {
      "type": "number",
      "title": "Storage",
      "default": 1.0
     }
    },
    "type": "object",
    "required": [
     "model_id"
    ],
    "title": "DecisionReq"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RegisterReq": {
    "properties": {
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "password": {
      "type": "string",
      "title": "Password"
     },
     "role": {
      "type": "string",
      "title": "Role",
      "default": "viewer"
     },
     "division": {
      "type": "string",
      "title": "Division",
      "default": "GODS"
     }
    },
    "type": "object",
    "required": [
     "email",
     "password"
    ],
    "title": "RegisterReq"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "auth/login"
     }
    }
   }
  }
 }
}